refactor: unify Logger, strip dead aggregate ctors, drop dead common.cpp
- Replace cppup::cli::Logger (virtual interface) with concept-based
  cppup::logger::Logger: single virtual log(LogLevel, string_view) +
  non-virtual info/warning/error/debug wrappers. ConsoleLogger now
  inherits from the polymorphic base and preserves the [LEVEL] format
  when category is empty. SilentLogger moved to logger/, deduplicating
  two test-local copies. set_verbose(bool) replaced with setGlobalConfig
  (LogLevel::Debug) at the two existing call sites.
- Drop forwarding-only ctors from Flag, Definition, PackageInfo,
  Toolchain, Module, Profile, Test, ResolvedPackage; they're now plain
  C++20 aggregates with P0960 paren-init at call sites. Added explicit
  '= {}' defaults on trailing members so -Wmissing-field-initializers
  stays quiet on aggregate-init call sites.
- Delete src/core/cli/commands/common.cpp (613 lines of scaffolding
  superseded by template-embedding in init.cpp; never reached from any
  caller). Keep common.h as transitive re-export header.
- Wrap build_step_executor.cpp helpers in an anonymous namespace; drop
  redundant 'static' qualifiers.
- Fix latent breakage: PackageImpl had been changed to a class inheriting
  privately from PackageInterface; restored public inheritance. Add
  defaulted ~CommandContext() and explicit copy/move=delete on
  ProcessRunner to satisfy Rule of Five.
- clang-tidy: tighten HeaderFilterRegex from '^.*/(src|include)/' to
  '^.*/cppup/(src|include)/' — the old regex matched /usr/include/gtest/
  by accident, surfacing gtest macro contexts in every test file's tidy
  output. Disable readability-redundant-member-init (conflicts with
  -Wmissing-field-initializers under -Werror). Broaden private/class
  member naming from snake_case_ to aNy_CasE_ so camelBack_ is accepted
  alongside the existing snake_case_ members.
- Promote LogLevel enum to std::uint8_t base.

CPPUP_SKIP_TIDY=1: this commit touches files that carry pre-existing tidy
warnings unrelated to the refactor. Cleaning those up is a separate pass.

+++ src/core/cli/cli_application.hpp
@@ -20,7 +20,7 @@ namespace cppup::cli
 class ErrorHandler
 {
  public:
-  enum class ErrorCode : int
+  enum class ErrorCode : uint8_t
   {
     Success          = 0,
     InvalidArguments = 1,
